New docker.types subpackage containing advanced config dictionary types
Tests and docs updated to match
docker.utils.types has been moved to docker.types
