+++ docker/api/__init__.py
@@ -5,9 +5,6 @@ from .daemon import DaemonApiMixin
 from .exec_api import ExecApiMixin
 from .image import ImageApiMixin
 from .network import NetworkApiMixin
-from .service import (
-    ServiceApiMixin, TaskTemplate, ContainerSpec, Mount, Resources,
-    RestartPolicy, UpdateConfig
-)
+from .service import ServiceApiMixin
 from .swarm import SwarmApiMixin
 from .volume import VolumeApiMixin
