+++ docker/utils/__init__.py
@@ -8,8 +8,6 @@ from .utils import (
     create_ipam_config, create_ipam_pool, parse_devices, normalize_links,
 )
 
-from .types import LogConfig, Ulimit
-from .types import (
-    SwarmExternalCA, SwarmSpec,
-)
+from ..types import LogConfig, Ulimit
+from ..types import SwarmExternalCA, SwarmSpec
 from .decorators import check_resource, minimum_version, update_headers
