+++ setup.py
@@ -36,7 +36,8 @@ setup(
     url='https://github.com/docker/docker-py/',
     packages=[
         'docker', 'docker.api', 'docker.auth', 'docker.transport',
-        'docker.utils', 'docker.utils.ports', 'docker.ssladapter'
+        'docker.utils', 'docker.utils.ports', 'docker.ssladapter',
+        'docker.types',
     ],
     install_requires=requirements,
     tests_require=test_requirements,
